requirements.txt
fastapi
uvicorn
openai
whisper
torch
pydantic
python-dotenv
